feat: neon bring-up fixes, generated api-types, sign-in UX polish
Backend:
- add spring-boot-starter-flyway (migrations never ran outside tests on Boot 4)
- sentry 8.46.0 spring-boot-4-starter (old starter fails introspection on Boot 4)
- DirectoryStudent.level short to match smallint schema
- verified live on Neon: 7 migrations applied, login -> directory -> session
  -> check-in -> summary -> admin report all green

Frontend:
- generated @kev/api-types from the live OpenAPI spec
- password show/hide eye toggle; outline lock icon matching the mail icon
- glass toast notifications (ToastHost + toast store) replace inline errors
- keyboard compresses the layout so Verify Account stays in view; inputs
  layered above the hero logo

+++ backend/src/main/java/com/kev/backend/directory/DirectoryStudent.java
@@ -33,8 +33,9 @@ public class DirectoryStudent {
     @Column(nullable = false)
     private String programme;
 
+    /** smallint in the schema (100..700). */
     @Column(nullable = false)
-    private int level;
+    private short level;
 
     @Column(name = "photo_url", nullable = false)
     private String photoUrl;
